+++ lib/postgres.dart
@@ -169,6 +169,12 @@ abstract class Session {
   });
 }
 
+/// A [Session] with transaction-related helper method(s).
+abstract class TxSession extends Session {
+  /// At the end of the transaction callback, the session will be closed with a `ROLLBACK`.
+  void rollback();
+}
+
 abstract class SessionExecutor {
   /// Obtains a [Session] capable of running statements and calls [fn] with
   /// it.
@@ -188,7 +194,7 @@ abstract class SessionExecutor {
   /// Note that other invocations on a [Connection] are blocked while a
   /// transaction is active.
   Future<R> runTx<R>(
-    Future<R> Function(Session session) fn, {
+    Future<R> Function(TxSession session) fn, {
     TransactionSettings? settings,
   });